Added a minimal Matrix module, move the README back to HTML.
The markdown README just wasn't rendering properly, so I moved it back to
HTML.

The matrix module is just a type wrapper, declaring that a tropical matrix is
just a normal matrix with tropical values.

I need to build in unit tests, probably with QuickCheck.

I also removed the equality requirement from tropicality, because it is implied
by the sortability requirement.